Refactor of main.py to make each step a modular function
