+++ requirements.txt
@@ -1,0 +1,1 @@
+google-cloud-bigquery==3.4.1
