+++ tweets.jsonl
@@ -1,10 +1,10 @@
-{"edit_history_tweet_ids": ["1611099457915097088"], "id": "1611099457915097088", "text": "New Year - New Job! ODI is accepting applications for our Principal Analytics Engineer position. If you love engaging with people and data, this might be the perfect opportunity for you!\n\nLearn more about ODI and this exciting opportunity at https://t.co/xNqI63RA8A #hiring #data https://t.co/TwqqkY7STJ"}
-{"edit_history_tweet_ids": ["1611035156122406912"], "id": "1611035156122406912", "text": ".@worldcoin is looking for Senior Analytics Engineer\n\n\ud83d\uddd3  Full-time\n\ud83d\udccd  Remote: Worldwide\n\nLearn more \ud83d\udc49 https://t.co/VZBderm6xd\n\n#web3 #blockchain #hiring #remotework"}
-{"edit_history_tweet_ids": ["1610988236481957888"], "id": "1610988236481957888", "text": "Kinsta is seeking a DevOps Engineer (Analytics) based in the Milan area. For more details, visit the listing. #atmosjobs #hiring #cloudjobs https://t.co/CQmMZ1oEx9"}
-{"edit_history_tweet_ids": ["1610946776260902912"], "id": "1610946776260902912", "text": "#Hiring for Data Engineer (Data Analytics) (IITs, BITS, NITs..(Tier1/ Tier2 colleges))\nLocation:Hyderabad\nExperience: 2+\nSalary: Upto 18 LPA\nShare your #resume to\nfalguni.kayasth@mantras2success.com\nOR\nClick to Apply\nhttps://t.co/aeB7GjBE88\n#Mantras2Success \n#dataanalytics #data"}
-{"edit_history_tweet_ids": ["1610652638479257600"], "id": "1610652638479257600", "text": "#hiring\n\nAnalytics Engineer https://t.co/c8tkDbYE8a"}
-{"edit_history_tweet_ids": ["1610509019420200960"], "id": "1610509019420200960", "text": "RT @Mantras2Success: #Hiring for Data Engineer (Data Analytics) (IITs, BITS, NITs..(Tier1/ Tier2 colleges))\nLocation: Hyderabad\nExperience:\u2026"}
-{"edit_history_tweet_ids": ["1610508879409864704"], "id": "1610508879409864704", "text": "#Hiring for Data Engineer (Data Analytics) (IITs, BITS, NITs..(Tier1/ Tier2 colleges))\nLocation: Hyderabad\nExperience: 2+\nSalary: Upto 18 LPA\nShare your #resume to\nfalguni.kayasth@mantras2success.com\nOR\nClick to Apply\nhttps://t.co/aeB7GjBE88\n#Mantras2Success \n#dataanalytics #data"}
-{"edit_history_tweet_ids": ["1610497501848653826"], "id": "1610497501848653826", "text": "Kinsta is seeking a DevOps Engineer (Analytics) based in the Milan area. For more details, visit the listing. #atmosjobs #hiring #cloudjobs https://t.co/CQmMZ1oEx9"}
-{"edit_history_tweet_ids": ["1610306029647953923"], "id": "1610306029647953923", "text": "Analytics Engineer\n\n\ud83c\udf0d ANYWHERE IN THE WORLD\n\ud83c\udfe2 Toggl\n\n#jobs #likeWFH #Hiring #NowHiring\n#remote #remotejobs #remotework\n#JobSearch #JobHunt #JobOpening\n\nhttps://t.co/ZMAH2O1TW8"}
-{"edit_history_tweet_ids": ["1610221117045891072"], "id": "1610221117045891072", "text": "RT @rustjobs_dev: \ud83d\udea8 Remote #Rust Job \ud83e\udd80\n\n@liquidanalytics is building next-generation analytics engines with Apache Arrow and Rust.\n\nThey ar\u2026"}
+{"edit_history_tweet_ids": ["1612546394556104704"], "id": "1612546394556104704", "text": "RT @AngelaBoudreau5: Harvard University Information Technology is #hiring a Business Intelligence Engineer. As a member of our reporting, A\u2026"}
+{"edit_history_tweet_ids": ["1612542160494010375"], "id": "1612542160494010375", "text": "Harvard University Information Technology is #hiring a Business Intelligence Engineer. As a member of our reporting, Analytics, and Integrated Data (RAID) Business Intelligence Engineering (BI Engineering) team, you will participate in software developmen\u2026https://t.co/FNlCuLxBhE"}
+{"edit_history_tweet_ids": ["1612517059631452161"], "id": "1612517059631452161", "text": "Kinsta is seeking a DevOps Engineer (Analytics) based in the Milan area. For more details, visit the listing. #atmosjobs #hiring #cloudjobs https://t.co/CQmMZ1oEx9"}
+{"edit_history_tweet_ids": ["1612396329694629889"], "id": "1612396329694629889", "text": "#Hiring for Data Engineer (Data Analytics) (IITs, BITS, NITs..(Tier1/ Tier2 colleges))\nLocation: Hyderabad\nExperience: 2+Years\nSalary: Upto 18 LPA\nShare your #resume to\nfalguni.kayasth@mantras2success.com\nOR\nClick to Apply\nhttps://t.co/aeB7GjBE88\n#Mantras2Success \n#dataanalytics"}
+{"edit_history_tweet_ids": ["1612086895755788289"], "id": "1612086895755788289", "text": "Metabase is hiring Analytics Engineer\n\n\ud83d\udd27 #sql\n\ud83c\udf0e Remote\n\u23f0 Full-time\n\ud83d\udcb0 $90k - $230k (Estimate)\n\ud83c\udfe2 Metabase\n\nJob details https://t.co/k47ECg8dSo\n#jobalert #jobsearch #hiring"}
+{"edit_history_tweet_ids": ["1611865886343954432"], "id": "1611865886343954432", "text": "RT @FrontendLA: #CreveCoeur #jobs #hiring #opportunity @ Bayer | Data Engineer (Expert) - Data &amp; Analytics - IT - Corp - US https://t.co/9N\u2026"}
+{"edit_history_tweet_ids": ["1611858019792613376"], "id": "1611858019792613376", "text": "#CreveCoeur #jobs #hiring #opportunity @ Bayer | Data Engineer (Expert) - Data &amp; Analytics - IT - Corp - US https://t.co/9NfEAaiXwP"}
+{"edit_history_tweet_ids": ["1611830031960915969"], "id": "1611830031960915969", "text": "Ready for a new exciting #job? @Chainalysis is #hiring a #Sr. DevOps Engineer, Advanced Analytics for #New York, Washington, DC. Help solve #cybercrime and make a difference in #crypto:   https://t.co/TuzZhPFotD\n\n#Jobsearch #Careers #Employment #Cryptojobs #Cryptocrime #Chain"}
+{"edit_history_tweet_ids": ["1611519665788047360"], "id": "1611519665788047360", "text": "#Huntsville #jobs #hiring #opportunity @ Gray Analytics | Network Engineer https://t.co/Tos6nmjn1Z"}
+{"edit_history_tweet_ids": ["1611422891467501568"], "id": "1611422891467501568", "text": "#Remote #jobs #hiring #opportunity @ V-Soft Consulting | Senior Web Analytics Engineer https://t.co/BMzSMOS4v7"}
